Clean up GD32 flash/PTP debug logging
Normalizes flash state enum names in the H7 flashcode path, tightens debug format usage with explicit casts, and keeps the GD32 include for IWYU. Also adds the missing GD32F450 9VEE (512KB) device mapping in gd32.json and introduces a new gd32_debug.h header.

+++ lib-gd32/device/enet/ptp/gd32_ptp.cpp
@@ -93,7 +93,7 @@ void Gd32PtpStart() {
     DEBUG_ENTRY();
     DEBUG_PRINTF("PTP_TICK=%u", gd32::ptp::kPtpTick);
     DEBUG_PRINTF("ADJ_FREQ_BASE_INCREMENT=%u", gd32::ptp::kAdjFreqBaseIncrement);
-    DEBUG_PRINTF("ADJ_FREQ_BASE_ADDEND=x%X", gd32::ptp::kAdjFreqBaseAddend);
+    DEBUG_PRINTF("ADJ_FREQ_BASE_ADDEND=x%X", static_cast<unsigned>(gd32::ptp::kAdjFreqBaseAddend));
 
     struct tm tmbuf;
     memset(&tmbuf, 0, sizeof(struct tm));
